+++ tests/Linna/Authentication/LoginAttemptTest.php
@@ -94,7 +94,7 @@ class LoginAttemptTest extends TestCase
         $loginAttempt->ipAddress = '192.168.1.2';
         $loginAttempt->when = \date('YmdHis', \time());
 
-        $this->assertSame(null, $loginAttempt->getId());
+        $this->assertSame(false, $loginAttempt->hasId());
 
         self::$enhancedAuthenticationMapper->save($loginAttempt);